pyproject.toml
[tool.poetry]
name = "bit2me-token"
version = "0.1.0"
description = ""
authors = ["seik <ivanmartinga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
eth-brownie = "^1.15.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pytest = "^6.2.4"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

